Added additional humanoid render layers
Added a layer for "belly" sprites to the Humanoid renderer. This allows rendering a custom animated sprite (same as body animation) per species, which is only visible when not wearing chest armor. Can be set via `player.setBellyImage(String)` and `npc.setBellyImage(String)`.
Added a layer for "groin" sprites to the Humanoid renderer. This allows rendering a custom animated sprite (same as body animations) per species, which is only visible when not wearing legs armor. Can be set via `player.setGroinImage(String)` and `npc.setGroinImage(String)`.